Remove random_velocity unit assumptions
Currently, the `random_velocity` function assumes the system masses are passed as `u` but if you use `g/mol` then this function will break whatever simulation you try to run. I have not actually fixed the code here it's just something I would like to fix at some point.

I'm not really sure the best way to handle this. Lately I've been running into units issues left and right. The main point of Unitful is so that the units are correct with minimal effort from the user, but I think it kind of defeats the point if I have to do a bunch of unit conversions before hand and declare all the unit types when making loggers or random velocities. At the moment it doesn't "just work".