Implements virial calculation on the GPU
Adds the calculation of the virial tensor to the GPU path.

Taking advantage of the functionality to store the force in per-atom buffers, this involves allocating additional equivalent memory to store intermediate results and updating the force kernels to accumulate the virial contribution from each pairwise interaction.

The final virial tensor is then transferred back to the CPU.

The modification to the AbstractGPUKernels is much more similar to what is done in the CPU path, storing the virial in a system-wide kernel.