Adds PBC correction to structure and traj writers
Adds the option to correct coordinates when writing structures, either keeping molecules whole (`:pbc`) or wrapping atoms individually inside the simulation box (`:wrap`).

This is controlled via the `correction` keyword in
`write_structure` and `TrajectoryWriter`.

Also fixed the VirialLogger to return sys.virial instead and store it in an appropriate matrix