New indexing in virial for specific interactions. Optimizations in the virial pipeline.
Still testing whether or not the virial is properly computed for specific interactions. Ensured that the reference atom for the virial is the same as the reference atom when calculatig 3 and 4 body interactions.

Now, the `forces!` function with and without the virial path are both compiled and called only when needed. That is, we only use the virial path in integration steps where the barostats requiring it are going to be called. Otherwise, the non virial path will be used, saving computation.