Virial removed from System struct
Changed the virial, kinetic energy and pressure tensors location from the System struct to inhabit a buffer struct. This buffer also carries the forces buffers.

This changed caused many rewrites in the code, all related to having to pass the buffer to the couplers, the loggers, as well as how the calculation of some magnitudes (kinetic energy, temperature, ...) is handled.