Better docs, assertions into errors and potential energy chenges
Better documented the MBAR pipeline, still need to add a couple of layers of abstraction for improved ease of use.

Changed all assertions into appropriate error throws to avoid possible optimisation level issues.

Small refactor on the potential energy calls. Now the buffers are only  created in the outermost abstraction layer and only if the array type is a GPU array, otherwise the buffers are Nothing