Added distance CVs and GPU tests
Added functionality to use various distances between two groups of atoms as a CV, incl. minimum and maximum distances and the distance between centres of mass. Added tests for these new distance functions.   Added tests of running biased simulations on the GPU. To avoid problems with scalar indexing in calculate_cv methods, calculations of energies and forces resulting from the bias in AtomsCalculators.potential_energy and AtomsCalculators.forces! are not performed on the GPU. Changed rmsd function to call kabsch function to avoid code duplication, and introduced the wrapper function ref_kabsch, which is marked as inactive to Enzyme (rather than kabsch itself).